LAVA-533 - split the help menu.
Create isolated V1 and V2 directories in doc.
Add support to settings.conf to turn on/off V1 and/or V2 docs.
Add support to settings.conf to add links to instance-specific docs.
Correct the indentation in templatetags, utils (pylint).

This change will need a small change to the packaging
once merged.

Change-Id: Ifb1ac40df968726cfb2c228d2c3398d9c0aea5af